Avoid firing event if no files changed
If the changes that triggered the event were only to known files, and
all of those files have the same hashes as when we began, we shouldn't
actually reload the command.

This optimization does *not* short circuit command reload if any of
these conditions hold:

- a changed file is larger than 20MB, e.g. changes to these files
always trigger a reload.

- more than 500 files changed

- a directory or an unknown file was changed.

Fixes jmhodges/justrun#23.